requirements.txt
psycopg2-binary
django
postgres
authlib ~= 1.0
python-dotenv ~= 0.19
requests ~= 2.27
django-cors-headers
djangorestframework
markdown
django-filter